Remove SourceConfig objects from classes that do not need it; minor improvements
