+++ quill-jdbc/src/main/scala/io/getquill/JdbcSourceConfig.scala
@@ -5,12 +5,12 @@ import java.util.Properties
 import com.zaxxer.hikari.{ HikariConfig, HikariDataSource }
 import io.getquill.naming.NamingStrategy
 import io.getquill.sources.SourceConfig
-import io.getquill.sources.jdbc.ConfigJdbcSource
+import io.getquill.sources.jdbc.ConnectionJdbcSource
 import io.getquill.sources.sql.idiom.SqlIdiom
 
 import scala.collection.JavaConversions._
 
-class JdbcSourceConfig[D <: SqlIdiom, N <: NamingStrategy](val name: String) extends SourceConfig[ConfigJdbcSource[D, N]] {
+class JdbcSourceConfig[D <: SqlIdiom, N <: NamingStrategy](val name: String) extends SourceConfig[ConnectionJdbcSource[D, N]] {
 
   def configProperties = {
     val p = new Properties
