+++ quill-jdbc/src/main/scala/io/getquill/sources/jdbc/ConnectionJdbcSource.scala
@@ -8,8 +8,12 @@ import io.getquill.sources.sql.idiom.SqlIdiom
 
 import scala.util.DynamicVariable
 
-class ConfigJdbcSource[D <: SqlIdiom, N <: NamingStrategy](config: JdbcSourceConfig[D, N])
-  extends JdbcSource[D, N](config) {
+class ConnectionJdbcSource[D <: SqlIdiom, N <: NamingStrategy](config: JdbcSourceConfig[D, N])
+  extends JdbcSource[D, N] {
+
+  protected val dataSource = config.dataSource
+
+  override def close = dataSource.close
 
   private val currentConnection = new DynamicVariable[Option[Connection]](Some(dataSource.getConnection))
